+++ packages/common/src/relay/calling/actions/SendDigitsAction.ts
@@ -1,0 +1,15 @@
+import BaseAction from './BaseAction'
+import SendDigits from '../components/SendDigits'
+import SendDigitsResult from '../results/SendDigitsResult'
+
+export default class SendDigitsAction extends BaseAction {
+
+  constructor(public component: SendDigits) {
+    super(component)
+  }
+
+  get result(): SendDigitsResult {
+    return new SendDigitsResult(this.component)
+  }
+
+}