+++ packages/common/src/relay/calling/results/SendDigitsResult.ts
@@ -1,0 +1,8 @@
+import BaseResult from './BaseResult'
+import SendDigits from '../components/SendDigits'
+
+export default class SendDigitsResult extends BaseResult {
+  constructor(public component: SendDigits) {
+    super(component)
+  }
+}